refactor: split frontend/src/data.rs by domain (#311)
The 1311-line `frontend/src/data.rs` had ~45 `pub` fetch wrappers and
was well over the ~800-line file-shape cap in
`.claude/rules/05-rust-style.md`. Split the per-domain wrappers out
into `frontend/src/data/{books,authors,series,tags,progress,auth}.rs`
and keep `data.rs` as the module root holding the shared transport
plumbing (`DataError`, mobile reqwest helpers, `token_store`,
`web_auth_state`, `note_server_fn_err`).

Pure file-shape change. Every function keeps its original logic and
existing `#[cfg(feature = …)]` gates, so the public surface
(`omnibus_frontend::data::*`) stays identical — callers continue to
import through the data module unchanged.

Closes #311
