+++ CLAUDE.md
@@ -85,7 +85,8 @@ migrations/         — numbered SQL migrations embedded via sqlx::migrate!
 
 ```
 lib.rs              — Route, App, styles, ScreenLayout (feature-gated); also owns the App-wide `CurrentUser` context (`use_current_user`) — a single `/api/auth/me` fetch on boot fills it, components gate `is_admin` off the cached signal instead of refetching per mount, and a `web_auth_state` subscription refills on login / clears on 401. Web/SSR only — mobile uses bearer tokens.
-data.rs             — Feature-gated data transport (mobile=reqwest, web/server=rpc); transport fns return Result<T, DataError> (thiserror enum: Network/Http/Decode/Unauthorized/Other) so the 401 path pattern-matches DataError::Unauthorized instead of re-inspecting raw status codes
+data.rs             — Module root for the feature-gated data transport (mobile=reqwest, web/server=rpc); owns the `DataError` thiserror enum (Network/Http/Decode/Unauthorized/Other), the mobile transport helpers (`ServerUrl`, `http_client`, `with_bearer`, `note_status`, `drain_error`, `client_kind`), `token_store` (mobile bearer-token store), `web_auth_state` (web 401 channel), and `note_server_fn_err` (web server-fn 401 mapper). Per-domain wrappers live under data/ and are re-exported here.
+data/               — Per-domain wrappers split out from data.rs: `books.rs` (settings/library/ebooks/search/overrides/worker), `authors.rs` (author CRUD + photo upload/scan/delete), `series.rs`, `tags.rs`, `progress.rs` (F2.1 reading/listening sync), `auth.rs` (login/register/logout/me + mobile bearer flow). Every wrapper keeps its existing `#[cfg(feature = …)]` gates and platform-agnostic signature.
 view_prefs.rs       — F1.3 per-library ViewPrefs/ViewFilters persistence for the landing page (sort/filter/facet state); web → localStorage keyed by library path, mobile → in-memory map (resets on cold launch), server (SSR) → defaults so first-hydration markup matches. Shape lives in omnibus-shared for a future server-backed endpoint
 reader_progress.rs  — F2.2 reading-position **cache-aside layer** for the F2.1 progress-sync endpoint: web → localStorage keyed by uuid, mobile → in-memory map, SSR → no-op. The reader reads this sync for instant first-paint then reconciles against `GET /api/progress/{uuid}` (server wins on success), and writes here AND fire-and-forget POSTs to `/api/rpc/progress` on every page turn.
 audiobook_progress.rs — F2.3 listening-position cache-aside layer, audio analogue of `reader_progress`. Stores seconds (float) keyed by uuid (`omn.listening::{uuid}`), plus a single device-wide playback-rate preference (`omn.listening.rate`). Same web/mobile/SSR split; the listen page reads the local value for instant first-paint, reconciles against `GET /api/progress/{uuid}?format=audio`, and writes locally + fire-and-forget POSTs to `/api/rpc/progress` on every 5 s of playback plus on pause.
